feat: Add support for sub keyEncoding in base & index sub
The test for the 4x index (an index of an index) is refactored to decode
the key in the callback to make it easier to `.get(key)`. This is a
matter of perspective as you could either want a secondary index to be
seen as indexing against the primary index's keys or as against the
initial key. This is up to the author of the index as leaving the key as
encoded would mean the value is stored in a double subencoded key for
the 4x sub. In the test however, it seems better to showcase it as being
4x the original value.