feat: Add `.update()` to `sub` to support awaiting index update
Previously the callback for updating the index fired at a later point
and there was no reliable way to ensure all updates had happened. Now
with `.update()` you can await for an update to happen. Because
`.update()` is debounced, it will let the index iterate through all
pending updates. If a callback takes a long time however, it will not
wait for that if there are previous iterations.