fix: Ensure autobase is ready before getting initial version
The base cannot be trusted to the user to have ready as this is "hidden
knowledege" that the user is likely to not do. Excluding this
flexibility is minor as many operations call ready in `autobase` and if
the user wants to recompute the view, they should do it with the
`version` option as it is assumed the index version has changed.