feat: Clarify the difference between `sub.update()` & `sub.drained()`
The difference between `sub.update()` & `sub.drained()` has been
clarified so that `.update()` is when any change has happened in the
index or if no changes are necessary then immediately return. Aka it now
updates the index.

`sub.drained()` resolves when the index is done processing changes. It
does this by waiting watchers to update and then await the SubIndex to
`.update()`.

As part of these changes the `drain` event was removed as `.drained()`
doesn't need it anymore.